move schemaNameMapped and getter
