+++ dbms/src/Databases/test/gtest_database.cpp
@@ -25,7 +25,6 @@
 #include <Storages/IManageableStorage.h>
 #include <Storages/IStorage.h>
 #include <Storages/MutableSupport.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TMTStorages.h>
 #include <Storages/registerStorages.h>
@@ -34,6 +33,8 @@
 #include <common/logger_useful.h>
 
 #include <optional>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
