+++ dbms/src/Debug/MockTiDB.h
@@ -15,12 +15,13 @@
 #pragma once
 
 #include <Storages/ColumnsDescription.h>
-#include <Storages/Transaction/SchemaGetter.h>
 #include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/Types.h>
 
 #include <atomic>
+
+#include <TIDB/Schema/SchemaGetter.h>
 
 namespace DB
 {
