+++ dbms/src/Interpreters/IDAsPathUpgrader.cpp
@@ -33,12 +33,13 @@
 #include <Poco/FileStream.h>
 #include <Storages/MutableSupport.h>
 #include <Storages/PathPool.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/TiDBSchemaSyncer.h>
 #include <common/logger_useful.h>
 #include <fmt/core.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
