+++ dbms/src/Storages/System/StorageSystemDTSegments.cpp
@@ -23,8 +23,9 @@
 #include <Storages/MutableSupport.h>
 #include <Storages/StorageDeltaMerge.h>
 #include <Storages/System/StorageSystemDTSegments.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/Types.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
