+++ dbms/src/Storages/System/StorageSystemDTTables.cpp
@@ -23,7 +23,8 @@
 #include <Storages/MutableSupport.h>
 #include <Storages/StorageDeltaMerge.h>
 #include <Storages/System/StorageSystemDTTables.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
