+++ dbms/src/Storages/System/StorageSystemDatabases.cpp
@@ -21,9 +21,10 @@
 #include <Databases/IDatabase.h>
 #include <Interpreters/Context.h>
 #include <Storages/System/StorageSystemDatabases.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/Types.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 
 namespace DB
