+++ dbms/src/Storages/System/StorageSystemTables.cpp
@@ -27,10 +27,11 @@
 #include <Storages/IManageableStorage.h>
 #include <Storages/MutableSupport.h>
 #include <Storages/System/StorageSystemTables.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/Types.h>
 #include <Storages/VirtualColumnUtils.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
