+++ dbms/src/Storages/Transaction/TiDB.cpp
@@ -23,8 +23,9 @@
 #include <Poco/StringTokenizer.h>
 #include <Storages/MutableSupport.h>
 #include <Storages/Transaction/Collator.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TiDB.h>
+
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
