+++ dbms/src/TIDB/Schema/SchemaBuilder.cpp
@@ -35,16 +35,16 @@
 #include <Parsers/parseQuery.h>
 #include <Storages/IManageableStorage.h>
 #include <Storages/MutableSupport.h>
-#include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TypeMapping.h>
+#include <TIDB/Schema/SchemaBuilder-internal.h>
+#include <TIDB/Schema/SchemaBuilder.h>
 #include <common/logger_useful.h>
 
 #include <boost/algorithm/string/join.hpp>
 #include <tuple>
 
-#include <TIDB/Schema/SchemaBuilder-internal.h>
-#include <TIDB/Schema/SchemaBuilder.h>
+#include <TIDB/Schema/SchemaNameMapper.h>
 
 namespace DB
 {
