+++ dbms/src/TIDB/Schema/SchemaBuilder.h
@@ -15,8 +15,9 @@
 #pragma once
 
 #include <Interpreters/Context.h>
-#include <Storages/Transaction/SchemaGetter.h>
 #include <Storages/Transaction/TMTStorages.h>
+
+#include <TIDB/Schema/SchemaGetter.h>
 
 namespace DB
 {
